# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Honey School API Gateway
  description: Public contract for the Honey School API Gateway.
  version: 0.1.0
servers:
- url: /api
  description: Honey School web-app API proxy
paths:
...
  /students/managed:
    post:
      tags:
      - Users
      summary: Create managed student account
      description: Creates or returns a teacher-managed Keycloak student account and
        app-level profile. Requires TEACHER or ADMIN role.
      operationId: createManagedStudent
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManagedStudentRequest"
        required: true
      responses:
        "200":
          description: Managed student profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfileResponse"
        "400":
          description: Invalid managed student payload
        "401":
          description: Missing or invalid bearer token
        "403":
          description: Current user is not a teacher/admin
        "409":
          description: Username or email already belongs to another account
      security:
      - bearerAuth: []
...
  /student-invites/consume:
    post:
      tags:
      - student-invite-controller
      operationId: consume
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInviteConsumeRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentInviteConsumeResponse"
  /api/student-invites/consume:
    post:
      tags:
      - student-invite-controller
      operationId: consume_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInviteConsumeRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentInviteConsumeResponse"
...
components:
  schemas:
...
    UserProfileResponse:
      type: object
      properties:
        subject:
          type: string
        username:
          type:
          - string
          - "null"
        email:
          type:
          - string
          - "null"
        name:
          type:
          - string
          - "null"
        roles:
          type: array
          items:
            type: string
        displayName:
          type:
          - string
          - "null"
        locale:
          type:
          - string
          - "null"
        countryCode:
          type:
          - string
          - "null"
        timezone:
          type:
          - string
          - "null"
        learningGoal:
          type:
          - string
          - "null"
        updatedAt:
          type: string
          format: date-time
        managedByTeacher:
          type: boolean
        birthDate:
          type:
          - string
          - "null"
          format: date
        lessonTranslationAllowed:
          type: boolean
          readOnly: true
      required:
      - lessonTranslationAllowed
      - managedByTeacher
      - roles
      - subject
      - updatedAt
...
    ManagedStudentRequest:
      type: object
      properties:
        username:
          type: string
          maxLength: 64
          minLength: 3
          pattern: "^[A-Za-z0-9._-]+$"
        firstName:
          type: string
          maxLength: 120
          minLength: 0
        lastName:
          type:
          - string
          - "null"
          maxLength: 120
          minLength: 0
        email:
          type:
          - string
          - "null"
          format: email
          maxLength: 320
          minLength: 0
      required:
      - firstName
      - username
...
    StudentInviteConsumeRequest:
      type: object
      properties:
        token:
          type: string
          maxLength: 255
          minLength: 0
      required:
      - token
    StudentInviteConsumeResponse:
      type: object
      properties:
        status:
          type: string
        accessToken:
          type:
          - string
          - "null"
        refreshToken:
          type:
          - string
          - "null"
        idToken:
          type:
          - string
          - "null"
        expiresIn:
          type:
          - integer
          - "null"
          format: int64
        continueUrl:
          type:
          - string
          - "null"
        opensAt:
          type:
          - string
          - "null"
          format: date-time
        scheduledStart:
          type:
          - string
          - "null"
          format: date-time
        scheduledEnd:
          type:
          - string
          - "null"
          format: date-time
        retryAfterSeconds:
          type:
          - integer
          - "null"
          format: int64
      required:
      - status
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT